+++ instrumentation/net/http/otelhttp/handler.go
@@ -109,23 +109,23 @@ func handleErr(err error) {
 func (h *middleware) createMeasures() {
 	var err error
 	h.requestBytesCounter, err = h.meter.Int64Counter(
-		RequestContentLength,
+		serverRequestSize,
 		metric.WithUnit("By"),
-		metric.WithDescription("Measures the size of HTTP request content length (uncompressed)"),
+		metric.WithDescription("Measures the size of HTTP request messages."),
 	)
 	handleErr(err)
 
 	h.responseBytesCounter, err = h.meter.Int64Counter(
-		ResponseContentLength,
+		serverResponseSize,
 		metric.WithUnit("By"),
-		metric.WithDescription("Measures the size of HTTP response content length (uncompressed)"),
+		metric.WithDescription("Measures the size of HTTP response messages."),
 	)
 	handleErr(err)
 
 	h.serverLatencyMeasure, err = h.meter.Float64Histogram(
-		ServerLatency,
+		serverDuration,
 		metric.WithUnit("ms"),
-		metric.WithDescription("Measures the duration of HTTP request handling"),
+		metric.WithDescription("Measures the duration of inbound HTTP requests."),
 	)
 	handleErr(err)
 }
@@ -225,7 +225,7 @@ func (h *middleware) serveHTTP(w http.ResponseWriter, r *http.Request, next http
 	span.SetStatus(semconv.ServerStatus(rww.statusCode))
 	span.SetAttributes(h.traceSemconv.TraceResponse(semconv.ResponseTelemetry{
 		StatusCode: rww.statusCode,
-		ReadBytes:  int(bw.read),
+		ReadBytes:  int(bw.read.Load()),
 		ReadError:  bw.err,
 		WriteBytes: int(rww.written),
 		WriteError: rww.err,
@@ -237,7 +237,7 @@ func (h *middleware) serveHTTP(w http.ResponseWriter, r *http.Request, next http
 		attributes = append(attributes, semconv.HTTPStatusCode(rww.statusCode))
 	}
 	o := metric.WithAttributes(attributes...)
-	h.requestBytesCounter.Add(ctx, bw.read, o)
+	h.requestBytesCounter.Add(ctx, bw.read.Load(), o)
 	h.responseBytesCounter.Add(ctx, rww.written, o)
 
 	// Use floating point division here for higher precision (instead of Millisecond method).